+++ salus_it600/const.py
