+++ salus_it600/exceptions.py
