+++ setup.py
@@ -2,7 +2,6 @@
 """The setup script."""
 import os
 import re
-import sys
 
 from setuptools import find_packages, setup
 
@@ -10,14 +9,13 @@ from setuptools import find_packages, setup
 def get_version():
     """Get current version from code."""
     regex = r"__version__\s=\s\"(?P<version>[\d\.]+?)\""
-    path = ("pyit600", "__version__.py")
+    path = ("salus_it600", "__version__.py")
     return re.search(regex, read(*path)).group("version")
 
 
 def read(*parts):
     """Read file."""
     filename = os.path.join(os.path.abspath(os.path.dirname(__file__)), *parts)
-    sys.stdout.write(filename)
     with open(filename, encoding="utf-8", mode="rt") as fp:
         return fp.read()
 
@@ -25,7 +23,7 @@ def read(*parts):
 with open("README.md") as readme_file:
     readme = readme_file.read()
 
-setup(author="Julius Vitkauskas",
+setup(author="Julius Vitkauskas, Jordi-14",
       author_email="zadintuvas@gmail.com",
       classifiers=[
           "Development Status :: 3 - Alpha",
@@ -33,17 +31,16 @@ setup(author="Julius Vitkauskas",
           "Programming Language :: Python :: 3.7",
           "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator"
       ],
-      description="Asynchronous Python client for Salus IT600 devices",
+      description="Asynchronous Python client for Salus iT600 devices",
       include_package_data=True,
       install_requires=["aiohttp>=3.8.1", "cryptography>=38.0.1"],
       keywords=["salus", "it600", "api", "async", "client"],
       license="MIT license",
       long_description_content_type="text/markdown",
       long_description=readme,
-      name="pyit600",
-      packages=find_packages(include=["pyit600"]),
-      test_suite="tests",
-      url="https://github.com/jvitkauskas/pyit600",
+      name="salus-it600-client",
+      packages=find_packages(include=["salus_it600", "salus_it600.*"]),
+      url="https://github.com/Jordi-14/salus-it600-client",
       version=get_version(),
       zip_safe=False,
       python_requires='>=3.7',
